Update to be currency
