--- src/Salary/SalaryConfig.php
@@ -1,13 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Engineered\Salary;
-
-use Gacela\Framework\AbstractConfig;
-
-/**
- * Copyright 2022 - The Customer Bureau - All Rights Reserved
- */
-final class SalaryConfig extends AbstractConfig
-{
-
-}
